Added conditional checks to migration script for fts_targets table presence
It ensures migrations can run successfully even if the Mroonga table is broken and needs to be recreated,
as detailed in Redmine Full Text Search documentation.

ref: https://github.com/clear-code/redmine_full_text_search?tab=readme-ov-file#mroonga

+++ db/migrate/20231210061330_add_created_at_to_fts_targets_with_index.rb
@@ -3,6 +3,8 @@ FullTextSearch::Migration
 
 class AddCreatedAtToFtsTargetsWithIndex < ActiveRecord::Migration[5.2]
   def up
+    return if !table_exists?(:fts_targets)
+
     add_column :fts_targets, :created_at, :timestamp
 
     ActiveRecord::Base.transaction do
@@ -39,6 +41,8 @@ class AddCreatedAtToFtsTargetsWithIndex < ActiveRecord::Migration[5.2]
   end
 
   def down
+    return if !table_exists?(:fts_targets)
+
     if Redmine::Database.mysql?
       remove_index :fts_targets, :created_at
     else
